增加SDL的IP控件和时间控件，移除GTK，整理代码。

+++ DuiLib/Utils/DuiString.cpp
@@ -1378,7 +1378,7 @@ namespace DuiLib
 				{
 					s.Remove(_T("0x"));
 					s.Remove(_T("0X"));
-					AddByte(s.HexToInt64());
+					AddByte((BYTE)s.HexToInt64());
 					s.Empty();
 				}
 			}
@@ -1388,7 +1388,7 @@ namespace DuiLib
 		{
 			s.Remove(_T("0x"));
 			s.Remove(_T("0X"));
-			AddByte(s.HexToInt64());
+			AddByte((BYTE)s.HexToInt64());
 			s.Empty();
 		}
 	}
